# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Workout Online",
        "description": "Документация для Workout",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/protected": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "description": "Метод для получения информации о текущем авторизованном пользователе или проверки авторизации",
                "operationId": "protected",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/type-of-workout": {
            "get": {
                "tags": [
                    "Type of workout"
                ],
                "description": "Метод возвращает все типы тренировок",
                "operationId": "typeOfWorkoutIndex",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "typeOfWorkout": {
                                            "$ref": "#/components/schemas/TypeOfWorkout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/type-of-workout/": {
            "post": {
                "tags": [
                    "Type of workout"
                ],
                "description": "Метод создает тип тренировки",
                "operationId": "typeOfWorkoutStore",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "test title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "test description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "typeOfWorkout": {
                                            "$ref": "#/components/schemas/TypeOfWorkout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/type-of-workout/{id}": {
            "get": {
                "tags": [
                    "Type of workout"
                ],
                "description": "Метод возвращает тип тренировки по id",
                "operationId": "typeOfWorkoutShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id типа тренировки",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "typeOfWorkout": {
                                            "$ref": "#/components/schemas/TypeOfWorkout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Type of workout"
                ],
                "description": "Метод обновляет тип тренировки",
                "operationId": "typeOfWorkoutUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id типа тренировки",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "test title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "test description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "typeOfWorkout": {
                                            "$ref": "#/components/schemas/TypeOfWorkout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Type of workout"
                ],
                "description": "Метод удаляет тип тренировки по id",
                "operationId": "typeOfWorkoutDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id типа тренировки",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "NOT FOUND"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "KindOfSport": {
                "title": "User",
                "description": "Пользователь",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Название вида спорта",
                        "example": "fitness"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "Profile": {
                "title": "Profile",
                "description": "Профиль пользователя",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "title": "first_name",
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Alexey"
                    },
                    "last_name": {
                        "title": "last_name",
                        "description": "Фамилия пользователя",
                        "type": "string",
                        "example": "Smirnov"
                    },
                    "age": {
                        "title": "age",
                        "description": "Возраст пользователя",
                        "format": "int64",
                        "example": 23
                    },
                    "gender": {
                        "title": "gender",
                        "description": "Пол пользователя",
                        "type": "string",
                        "example": "male"
                    },
                    "sports_experience": {
                        "title": "sports_experience",
                        "description": "Стаж тренировок пользователя",
                        "type": "string",
                        "example": "4 года"
                    },
                    "sports_level": {
                        "title": "sports_experience",
                        "description": "Спортивный опыт пользователя",
                        "type": "string",
                        "example": "advanced"
                    },
                    "kind_of_sport": {
                        "$ref": "#/components/schemas/KindOfSport"
                    },
                    "food_type": {
                        "title": "food_type",
                        "description": "Режим питания пользователя",
                        "type": "string",
                        "example": "cut"
                    },
                    "training_days": {
                        "title": "training_days",
                        "description": "Расписание тренировок пользователя",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "monday",
                            "wednesday",
                            "friday"
                        ]
                    },
                    "work_experience": {
                        "title": "work_experience",
                        "description": "Стаж работы тренера",
                        "type": "string",
                        "example": "1 год"
                    },
                    "rate_per_hour": {
                        "title": "rate_per_hour",
                        "description": "Ставка тренера за час",
                        "type": "number",
                        "format": "float",
                        "example": "1000"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Profile"
                }
            },
            "TypeOfWorkout": {
                "title": "TypeOfWorkout",
                "description": "Тип тренировки",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Название типа тренировки",
                        "type": "string",
                        "example": "power"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TypeOfWorkout"
                }
            },
            "User": {
                "title": "User",
                "description": "Пользователь",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "int64",
                        "example": 1
                    },
                    "email": {
                        "title": "title",
                        "description": "email пользователя",
                        "type": "string",
                        "example": "example@gmail.com"
                    },
                    "role": {
                        "title": "role",
                        "description": "Роль пользователя",
                        "type": "string",
                        "example": "trainer"
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}